Remove Visitor impl for TestNodeBuilder
